# legacy-codebase/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyDreamCampus API
  description: |
    Üniversite yönetim sistemi API dokümantasyonu.

    Tüm API'ler Traefik gateway üzerinden `http://localhost/api/...` şeklinde erişilir.

    ## Kimlik Doğrulama
    - `POST /api/auth/login` ile JWT token alınır
    - Token `Authorization: Bearer <token>` header'ı ile gönderilir
    - Traefik forward-auth middleware'i token'ı doğrular ve `X-User-ID`, `X-User-Role`, `X-User-Department` header'larını backend servislere iletir
    - Refresh token HttpOnly cookie olarak saklanır

    ## Roller
    - **admin**: Tam yetki
    - **teacher**: Ders, yoklama, not yönetimi
    - **student**: Kayıt, yoklama, not görüntüleme

    ## Rate Limiting
    Tüm endpoint'ler rate limiting ile korunur. Yanıt header'ları:
    - `X-RateLimit-Limit`: Zaman penceresi başına izin verilen maksimum istek sayısı
    - `X-RateLimit-Remaining`: Kalan istek sayısı
    - `X-RateLimit-Reset`: Limit sıfırlanma zamanı (Unix epoch saniye)

    Limit aşıldığında `429 Too Many Requests` döner.

    ## Hata Yönetimi
    Tüm endpoint'ler aşağıdaki genel hata yanıtlarını dönebilir:
    - `401 Unauthorized`: Geçersiz veya eksik JWT token
    - `429 Too Many Requests`: Rate limit aşıldı
    - `500 Internal Server Error`: Beklenmeyen sunucu hatası

    Hata yanıtları her zaman `ErrorResponse` formatındadır.

    ## Sağlık Kontrolü
    Her servis `GET /health` endpoint'i sunar. Bu endpoint'ler kimlik doğrulama gerektirmez.
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/grades/course/{course_id}/scores:
    post:
      operationId: submitScore
      tags: [Grades]
      summary: Not gir
      description: teacher veya admin rolü gerektirir. Tek öğrenci için not girişi.
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreRequest"
      responses:
        "200":
          description: Not girildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitScoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/grades/course/{course_id}/scores/bulk:
    post:
      operationId: bulkSubmitScores
      tags: [Grades]
      summary: Toplu not gir
      description: teacher veya admin rolü gerektirir. Birden fazla öğrenci için aynı assessment'a not girişi.
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkSubmitScoresRequest"
      responses:
        "200":
          description: Toplu not girişi sonucu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkSubmitScoresResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/grades/transcript/{student_id}:
    get:
      operationId: getTranscript
      tags: [Grades]
      summary: Transkript
      description: |
        Öğrenci kendi transkriptini, öğretmen danışan öğrencisinin transkriptini,
        admin herkesin transkriptini görüntüleyebilir.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Transkript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/grades/admin/appeal:
    post:
      operationId: processAppeal
      tags: [Grades]
      summary: Not itirazı işle
      description: Sadece admin. Kesinleşmiş notu değiştirir.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealScoreRequest"
      responses:
        "200":
          description: İtiraz sonucu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppealScoreResponse"
  /api/grades/admin/scores/unlock:
    post:
      operationId: unlockScore
      tags: [Grades]
      summary: Notu kilidi aç
      description: Sadece admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreLockRequest"
      responses:
        "200":
          description: Kilit açıldı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/grades/admin/scores/lock:
    post:
      operationId: lockScore
      tags: [Grades]
      summary: Notu kilitle
      description: Sadece admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreLockRequest"
      responses:
        "200":
          description: Kilitlendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ============================================================
  # MEAL SERVICE
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `POST /api/auth/login` ile alınır."
...
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Yetkisiz erişim
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Kayıt bulunamadı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Beklenmeyen sunucu hatası
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "INTERNAL_ERROR"
            error: "INTERNAL_ERROR"
            message: "Beklenmeyen bir hata oluştu"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email alanı gereklidir"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "işlem başarılı"
...
    SubmitScoreRequest:
      type: object
      required: [registration_id, slug]
      properties:
        registration_id:
          type: string
          format: uuid
        slug:
          type: string
          example: "midterm"
        score:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 100
        is_absent:
          type: boolean
          default: false
    BulkScoreEntry:
      type: object
      required: [registration_id]
      properties:
        registration_id:
          type: string
          format: uuid
        score:
          type: number
          format: float
          nullable: true
        is_absent:
          type: boolean
          default: false
    BulkSubmitScoresRequest:
      type: object
      required: [slug, scores]
      properties:
        slug:
          type: string
        scores:
          type: array
          items:
            $ref: "#/components/schemas/BulkScoreEntry"
    FinalizeResult:
      type: object
      properties:
        grading_type:
          type: string
        class_mean:
          type: number
          format: float
        total_students:
          type: integer
        passing_count:
          type: integer
        failing_count:
          type: integer
        attendance_failed_count:
          type: integer
    SubmitScoreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_number:
          type: string
        slug:
          type: string
        score:
          type: number
          format: float
          nullable: true
        is_absent:
          type: boolean
        graded_at:
          type: string
          format: date-time
        auto_finalized:
          type: boolean
        finalize_result:
          $ref: "#/components/schemas/FinalizeResult"
    BulkSubmitScoresResponse:
      type: object
      properties:
        slug:
          type: string
        success_count:
          type: integer
        auto_finalized:
          type: boolean
        finalize_result:
          $ref: "#/components/schemas/FinalizeResult"
...
    TranscriptResponse:
      type: object
      properties:
        student:
          type: object
          properties:
            student_number:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            department:
              type: string
            enrollment_year:
              type: integer
        semesters:
          type: array
          items:
            type: object
            properties:
              semester:
                type: string
              semester_display:
                type: string
              courses:
                type: array
                items:
                  type: object
                  properties:
                    course_code:
                      type: string
                    course_name:
                      type: string
                    credits:
                      type: integer
                    grade_point:
                      type: string
              semester_credits:
                type: integer
              semester_gpa:
                type: number
                format: float
        summary:
          type: object
          properties:
            total_credits:
              type: integer
            cumulative_gpa:
              type: number
              format: float
        generated_at:
          type: string
          format: date-time
    AppealScoreRequest:
      type: object
      required: [student_id, course_id, slug, new_score, reason]
      properties:
        student_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        slug:
          type: string
        new_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        reason:
          type: string
          minLength: 10
    AppealScoreResponse:
      type: object
      properties:
        student_id:
          type: string
          format: uuid
        course_code:
          type: string
        slug:
          type: string
        old_score:
          type: number
          format: float
        new_score:
          type: number
          format: float
        old_weighted_average:
          type: number
          format: float
        new_weighted_average:
          type: number
          format: float
        old_grade_point:
          type: string
        new_grade_point:
          type: string
        grading_type:
          type: string
        frozen_class_mean:
          type: number
          format: float
        frozen_class_stddev:
          type: number
          format: float
    ScoreLockRequest:
      type: object
      required: [registration_id, slug]
      properties:
        registration_id:
          type: string
          format: uuid
        slug:
          type: string

    # ---- Meal ----
...